+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to the "claw" and/or "openclaw" extension will be documented in this file.
 
 ## Releases
+- 0.1.3
+  - new terminal session for ggc oc
+- 0.1.2
+  - magnet icon and theme
+- 0.1.1
+  - docs refinement
 - 0.1.0
   - use ggc oc for responsive tui
 - 0.0.9